Broke out hyperlinks extraction command
